requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
rembg[cpu]==2.0.59
onnxruntime==1.19.0
python-dotenv==1.0.1
